all: use runtime.IsOpenharmony instead of runtime.GOOS=="openharmony"
Change the value of runtime.GOOS from "openharmony" to "linux" when target os is openharmony,so that we don't need to change `runtime.GOOS == "linux"` to `runtime.GOOS == "linux" || runtime.GOOS == "openharmony"` for tens of thousands of existing Go open source projects.

Add runtime.IsOpenharmony to specify openharmony.

For #IBFOQM

Signed-off-by: WuGuangyao <wuguangyao@huawei.com>